backend/functions/requirements.txt
google-cloud-vision
functions-framework